refactor delete() to not create aggregate
previously an aggregate was created and then only used for lookup of a
single id before being dropped. Refactored this to be more sane and use
the known id directly.

Creating an aggregate may also fail because entities are not directly
linked in the db-model, e.g. creating an observationData based on
getEntity() will fail as `data.getFeature` is not populated by mappings.